Dev -> main 작품 메인으로 전달

+++ src/main/java/com/lit_map_BackEnd/domain/work/service/WorkServiceImpl.java
@@ -12,7 +12,9 @@ import com.lit_map_BackEnd.domain.character.service.CastService;
 import com.lit_map_BackEnd.domain.genre.entity.Genre;
 import com.lit_map_BackEnd.domain.genre.service.GenreService;
 import com.lit_map_BackEnd.domain.member.entity.Member;
+import com.lit_map_BackEnd.domain.member.entity.MemberRoleStatus;
 import com.lit_map_BackEnd.domain.member.repository.MemberRepository;
+import com.lit_map_BackEnd.domain.relation.service.RelationService;
 import com.lit_map_BackEnd.domain.work.dto.VersionListDto;
 import com.lit_map_BackEnd.domain.work.dto.VersionResponseDto;
 import com.lit_map_BackEnd.domain.work.dto.WorkRequestDto;
@@ -30,6 +32,8 @@ import org.springframework.validation.annotation.Validated;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -47,7 +51,7 @@ public class WorkServiceImpl implements WorkService{
     private final CastService castService;
     private final VersionService versionService;
     private final MemberRepository memberRepository;
-
+    private final RelationService relationService;
 
     @Autowired
     private final YoutubeService youtubeService;
@@ -60,7 +64,7 @@ public class WorkServiceImpl implements WorkService{
 
     @Override
     @Transactional
-    public int saveWork(@Valid WorkRequestDto workRequestDto) {
+    public int saveWork(Member member, @Valid WorkRequestDto workRequestDto) {
         // 기존에 존재하는 작품인지 아닌지 확인
         boolean isNew = false;
 
@@ -68,8 +72,16 @@ public class WorkServiceImpl implements WorkService{
         Version version = null;
 
         // 회원 증명
-        Member member = memberRepository.findById(workRequestDto.getMemberId())
-                .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.USER_NOT_FOUND));
+        MemberRoleStatus memberRoleStatus = member.getMemberRoleStatus();
+        Set<MemberRoleStatus> invalidStatuses = Set.of(
+                MemberRoleStatus.PENDING_MEMBER,
+                MemberRoleStatus.UNKNOWN_MEMBER,
+                MemberRoleStatus.WITHDRAWN_MEMBER
+        );
+
+        if (invalidStatuses.contains(memberRoleStatus)) {
+            throw new BusinessExceptionHandler(ErrorCode.WRITER_INVALID);
+        }
 
         // 이미 존재하는 작품이 있다면 기존의 작품 불러오기
         if(workRequestDto.getWorkId() != null) {
@@ -85,7 +97,6 @@ public class WorkServiceImpl implements WorkService{
                     .member(member)
                     .publisher(member.getPublisher())
                     .publisherDate(workRequestDto.getPublisherDate())
-                    .category(null)
                     .view(0)
                     .build();
             isNew = true;
@@ -205,7 +216,7 @@ public class WorkServiceImpl implements WorkService{
         Work work = workRepository.findById(workId)
                 .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.WORK_NOT_FOUND));
 
-        WorkResponseDto workData = getWorkData(workId, 0.1);
+        WorkResponseDto workData = getWorkData(workId, 1.0);
 
         List<VersionListDto> maps = versionService.versionList(work);
 
@@ -222,10 +233,15 @@ public class WorkServiceImpl implements WorkService{
 
     @Override
     @Transactional
-    public void deleteWork(Long workId) {
-        workRepository.findById(workId)
+    public void deleteWork(Member member, Long workId) {
+        Work work = workRepository.findById(workId)
                 .orElseThrow(() -> new BusinessExceptionHandler(ErrorCode.WORK_NOT_FOUND));
-        workRepository.deleteById(workId);
+
+        if (member.getId().equals(work.getMember().getId()) || member.getMemberRoleStatus() == MemberRoleStatus.ADMIN) {
+            workRepository.deleteById(workId);
+        } else {
+            throw new BusinessExceptionHandler(ErrorCode.WRITER_WRONG);
+        }
     }
 
     // 제출(true)과 수정/임시저장(false)을 구분하는 메소드
@@ -270,6 +286,9 @@ public class WorkServiceImpl implements WorkService{
 
         // 이는 초기 로딩 속도와 네트워크 효율성을 고려하여 제작
         VersionResponseDto version = versionService.findVersionByWorkAndNumber(work.getId(), versionNum);
+
+        //연관 작품 가져오기
+        // List<Work> recommendedWorks = relationService.recommendRelatedWorks(work);
 
         // Youtube 정보 조회
         String workTitle = work.getTitle(); // 작품의 제목 가져오기
